create utils_test.go file to avoid confusion with file names and what is being tested in what file, drop internal-only `PrintContextInternals` fn impacting test coverage, add test for `TruncateStringByBytes`
